API Funzionanti (FINALMENTE)

+++ Planner3.Data/REPO/ExerciseRepo.cs
@@ -16,30 +16,46 @@ namespace Planner3.Data.REPO
         {
             _ctx = ctx;
         }
-        public async Task<IEnumerable<Exercise>> GetExercisesAsync()
-        {
-            var exercises = await _ctx.Exercises.ToListAsync();
-            return exercises;
-        }
-        public async Task<Exercise> GetExercisesByIdAsync(int id)
-        {
-            return await _ctx.Exercises.FindAsync(id);
-        }
         public async Task<Exercise> CreateExerciseAsync(Exercise exercise)
         {
             _ctx.Exercises.Add(exercise);
-            _ctx.SaveChanges();
+            await _ctx.SaveChangesAsync();
             return exercise;
+        }
+        public async Task<List<Exercise>> GetExercisesAsync()
+        {
+            return await _ctx.Exercises
+                .Include(e => e.WorkoutSessions)
+                .ToListAsync();
+        }
+        public async Task<Exercise> GetExerciseByIdAsync(int id)
+        {
+            return await _ctx.Exercises
+                .Include(e => e.WorkoutSessions)
+                .FirstOrDefaultAsync(e => e.Id == id);
         }
         public async Task UpdateExerciseAsync(Exercise exercise)
         {
-            _ctx.Exercises.Update(exercise);
-            await _ctx.SaveChangesAsync();
+            var existingExercise = await _ctx.Exercises.FindAsync(exercise.Id);
+            if (existingExercise != null)
+            {
+                existingExercise.Name = exercise.Name;
+                existingExercise.Description = exercise.Description;
+                existingExercise.Type = exercise.Type;
+                existingExercise.MuscleGroup = exercise.MuscleGroup;
+
+                _ctx.Exercises.Update(existingExercise);
+                await _ctx.SaveChangesAsync();
+            }
         }
-        public async Task DeleteExerciseAsync(Exercise exercise)
+        public async Task DeleteExerciseAsync(int id)
         {
-            _ctx.Exercises.Remove(exercise);
-            await _ctx.SaveChangesAsync();
+            var exercise = await _ctx.Exercises.FindAsync(id);
+            if (exercise != null)
+            {
+                _ctx.Exercises.Remove(exercise);
+                await _ctx.SaveChangesAsync();
+            }
         }
     }
 }
