+++ Planner3.Data/REPO/UserRepo.cs
@@ -16,30 +16,47 @@ namespace Planner3.Data.REPO
         {
             _ctx = ctx;
         }
-        public async Task<IEnumerable<User>> GetUsersAsync()
-        {
-            var users = await _ctx.Users.ToListAsync();
-            return users;
-        }
-        public async Task<User> GetUsersByIdAsync(int id)
-        {
-            return await _ctx.Users.FindAsync(id);
-        }
+
+        // per la registrazione ->
         public async Task<User> CreateUserAsync(User user)
         {
             _ctx.Users.Add(user);
             await _ctx.SaveChangesAsync();
             return user;
         }
+        public async Task<User> GetUserByIdAsync(int id)
+        {
+            return await _ctx.Users.FindAsync(id);
+        }
+
+        // per il login ->
+        public async Task<User> GetUserByEmail(string email, string password)
+        {
+            return await _ctx.Users.FirstOrDefaultAsync(u => u.Email == email && u.Password == password);
+        }
         public async Task UpdateUserAsync(User user)
         {
-            _ctx.Users.Update(user);
-            await _ctx.SaveChangesAsync();
+            var existingUser = await _ctx.Users.FindAsync(user.Id);
+            if (existingUser != null)
+            {
+                existingUser.Weight = user.Weight;
+                existingUser.Height = user.Height;
+                existingUser.ExperienceLevel = user.ExperienceLevel;
+                existingUser.Goals = user.Goals;
+                existingUser.Name = user.Name;
+
+                _ctx.Users.Update(existingUser);
+                await _ctx.SaveChangesAsync();
+            }
         }
-        public async Task DeleteUserAsync(User user)
+        public async Task DeleteUserAsync(int id)
         {
-            _ctx.Users.Remove(user);
-            await _ctx.SaveChangesAsync();
+            var user = await _ctx.Users.FindAsync(id);
+            if (user != null)
+            {
+                _ctx.Users.Remove(user);
+                await _ctx.SaveChangesAsync();
+            }
         }
     }
 }
